move DOM translation functions into observer
